+++ lithops/storage/backends/gcp_storage/gcp_storage.py
@@ -15,6 +15,8 @@
 #
 
 import re
+import os
+import shutil
 import time
 import logging
 from requests.exceptions import SSLError as TooManyConnectionsError
@@ -85,6 +87,52 @@ class GCPStorageBackend:
             return stream
         else:
             return blob.download_as_string(start=start, end=end)
+
+    def upload_file(self, file_name, bucket, key=None, extra_args={}):
+        """Upload a file
+
+        :param file_name: File to upload
+        :param bucket: Bucket to upload to
+        :param key: object name. If not specified then file_name is used
+        :return: True if file was uploaded, else False
+        """
+        # If S3 key was not specified, use file_name
+        if key is None:
+            key = os.path.basename(file_name)
+
+        # Upload the file
+        try:
+            with open(file_name, 'rb') as in_file:
+                self.put_object(bucket, key, in_file)
+        except Exception as e:
+            logging.error(e)
+            return False
+        return True
+
+    def download_file(self, bucket, key, file_name=None, extra_args={}):
+        """Download a file
+
+        :param bucket: Bucket to download from
+        :param key: object name. If not specified then file_name is used
+        :param file_name: File to upload
+        :return: True if file was downloaded, else False
+        """
+        # If file_name was not specified, use S3 key
+        if file_name is None:
+            file_name = key
+
+        # Download the file
+        try:
+            dirname = os.path.dirname(file_name)
+            if dirname and not os.path.exists(dirname):
+                os.makedirs(dirname)
+            with open(file_name, 'wb') as out:
+                data_stream = self.get_object(bucket, key, stream=True)
+                shutil.copyfileobj(data_stream, out)
+        except Exception as e:
+            logging.error(e)
+            return False
+        return True
 
     def head_object(self, bucket_name, key):
         try: